Actually fixes `useClickOutside` bug. Refactors `Header` component to be more readable by breaking into smaller chunks. Fixes `StateScreen` background color.
