+++ src/components/StateScreens/StateScreen.js
@@ -5,8 +5,10 @@ import classes from './styles.module.scss';
 export default function StateScreen({ iconComponent: IconComponent, children, ...otherProps }) {
   return (
     <div className={classes['state-screen']} {...otherProps}>
-      <IconComponent className={classes['icon']} />
-      {children}
+      <div className={classes['state-screen-content']}>
+        <IconComponent className={classes['icon']} />
+        {children}
+      </div>
     </div>
   );
 }
